requirements.txt
flask
pymysql
pymongo
pyrebase4
python-dotenv
flask_jwt_extended
flask-cors